added status classes: BaseClass, WorkingStatus, TimedWorkingStatus (Status.h, Status.cpp) soon to be integrated into the different modules getElapsedMs(): now returns a float with more precision, can now accept an end time to use instead of now()

+++ RealBloom/src/RealBloom/Dispersion.cpp
@@ -343,7 +343,7 @@ namespace RealBloom
 
         if (m_state.working)
         {
-            float elapsedSec = (float)getElapsedMs(m_state.timeStart) / 1000.0f;
+            float elapsedSec = getElapsedMs(m_state.timeStart) / 1000.0f;
 
             if (m_capturedParams.methodInfo.method == DispersionMethod::CPU)
             {
